requirements.txt
hammock
requests
Pillow
-e git+https://github.com/leofiore/tilify#egg=tilify
